requirements.txt
Flask==2.3.2
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.0.1
Werkzeug==3.0.0
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.25
python-dotenv==1.0.1
